+++ src/lib.rs
@@ -67,11 +67,11 @@
 //!
 //! // Generate the posterior distribution P(θ|x); the distribution of
 //! // probable coin weights
-//! let posterior: Beta = prior.posterior(&obs);
+//! let posterior: Beta = prior.posterior(obs);
 //!
 //! // What is the probability that the next flip would come up heads
 //! // (true) given the observed flips (posterior predictive)?
-//! let p_heads = prior.pp(&true, &obs);
+//! let p_heads = prior.pp(&true, obs);
 //! ```
 
 #![cfg_attr(coverage_nightly, feature(coverage_attribute))]